:rocket:

+++ src/main/java/org/example/command/RemoveGreater.java
@@ -29,6 +29,10 @@ public class RemoveGreater extends Command {
                     ", получено: " + args.length;
         }
 
+        if (object == null){
+            return "Неверные аргументы";
+        }
+
         LabWork labWork = (LabWork) object;
 
         boolean flag = labWorkService.removeGreater(labWork);
